Bot will self re-run if crush
Bot will now stay in the raid list 1/3 of the time in-case user wants to see if all attacks sent properly.
In-case of exception, it will check current url and re-log to resume itself.
Now it will also go to the village rather than resource page, because of incoming attacks cause web to re-load itself